Thinning typeclass in order to derive most instances automatically
This required moving {or, and, nor, nand, any, all} out of the
typeclass. I regard it as exceedingly unlikely that anyone would wish
to redefine them anyway.

"Deriving via" is only available from GHC8.6.1; we give instances
explicitly in this library in order to compile with GHC7.0 and
upwards, but users defining their own instances with later versions
may be glad of the functionality.